+++ changelog.md
@@ -7,7 +7,7 @@ The formatting is based on *[Keep a Changelog](https://keepachangelog.com/en/1.0
 and this project adheres to *[Semantic Versioning](https://semver.org/spec/v2.0.0.html)*.
 
 ## [Unreleased]
-- Added ShuffleAllToDeck(), changelog.md, changelogtemplate.md / Nathan 2023-10-27
+- Added ShuffleAllToDeck(), Player class, changelog.md, changelogtemplate.md / Nathan 2023-10-27
 - Change file structure to split Card.py with card related classes / Nathan 2023-10-27
 - Added game phases / Nathan 2023-10-26
 - Changed flow of game loop / Nathan 2023-10-26